infer/requirements.txt
numpy
tensorflow
keras
azure.servicebus
pymongo
azure.storage.blob
PIL